Create Graph.R
This program requires R to run it. It requires the result from the name.pl program, Totals.java program, and the gene trees. It will use the taxa names from name.pl program and gene trees to count the number of taxa in gene trees and then find the percent of missing data.  Moreover, it will use the result from Totals.java to figure out the taxa level.  As the result, this program will use taxa level and percent of missing data to plot a graph.